gameplus/requirements.txt
Django==1.8
dj-database-url==0.3.0
dj-static==0.0.6
gunicorn==19.3.0
psycopg2==2.6
static3==0.5.1
itsdangerous==0.24
django-social-share==1.1.0
whitenoise==3.3.0
django-cors-headers==2.0.2
django-cors-middleware==1.3.1
Pillow==4.0.0